Unfortunately I got called into a different L1 discussion so I am late getting a plan to you.
1) We have a new 5.02 TeV MB sample that you should use, please see the email I sent to the discussion list.
2) We should show tomorrow the following item:

For the single region, 2x2, and 3x3 seeds, we should show the turn-on curve that reaches 100% efficiency at photon pt = 60 GeV all on the same plot and state the rates. I can make the combo plot and I have the single region turn-ons, so from you guys I need you to tell me where I can find the histogram file that contains the 2x2 and 3x3 turn-on curves, and which turn-on curve is the first to be 100% efficient at photon pt = 60GeV (using the photon-triggered data sample). I would then need you to tell me the rate of that threshold on the 5.02 TeV sample.